Fixed minor bugs
Fixed the following bugs:

The dropdown of streaming options showed incorrect info if no platform offered that specific item. Now the dropdown is empty.

Adding a favourite when the list of favourits was empty could launch an error due to not accounting for the particular case when the list became empty. Now works without issues.